# docs/public_api_reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "beetkeeper",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": ""
  },
  "paths": {
...
    "/api/import": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Start Import",
        "description": "Enqueue an import of the given paths and return the created (PENDING) job.\n\nSet `quiet=true` to import non-interactively (the `beet import -q` equivalent): no decision prompts.\nThe optional per-job settings (`quiet`, `logpath`, `group_albums`, `flat`, `set_fields`) default to the\ncorresponding beets config values when left unspecified.",
        "operationId": "start_import_api_import_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "List Imports",
        "description": "List one page of known import jobs, newest first (so page 1 shows the most recent submissions).",
        "operationId": "list_imports_api_import_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "The page number, 1-indexed.",
              "default": 1,
              "title": "Page"
            },
            "description": "The page number, 1-indexed."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum results per page.",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Maximum results per page."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportJob"
                  },
                  "title": "Response List Imports Api Import Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/import/clean_slate": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Start Clean Slate",
        "description": "Enqueue a clean-slate import: remove the named library entry, then import its source folder afresh.\n\nThe preview runs first: a blocking error is a 422, a source with fewer audio files than the entry has on\ndisk is a 409 until `allow_fewer_files` is set, and an unknown beets id is a 404. The job is then\ntracked like any other import (poll it, answer its decisions, abort it via the routes below); the worker\nre-runs the preview as its guard right before removing anything.",
        "operationId": "start_clean_slate_api_import_clean_slate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanSlateSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/import/{job_id}": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Get Import",
        "description": "Return a single import job (poll this for status / the pending decision).",
        "operationId": "get_import_api_import__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/{job_id}/decision": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Decide Import",
        "description": "Answer the decision an import is parked on; 409 if it isn't awaiting one.",
        "operationId": "decide_import_api_import__job_id__decision_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/{job_id}/abort": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Abort Import",
        "description": "Request cooperative cancellation of an in-flight import.",
        "operationId": "abort_import_api_import__job_id__abort_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "DecisionRequest": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ImportCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/ImportAction"
            },
            "type": "array",
            "title": "Allowed Actions"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "task_id",
          "prompt"
        ],
        "title": "DecisionRequest",
        "description": "Published when the importer needs a user choice for a task; the matching `ImportDecision` unblocks it."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportAction": {
        "type": "string",
        "enum": [
          "apply",
          "asis",
          "skip"
        ],
        "title": "ImportAction",
        "description": "beetkeeper-side mirror of beets' import `action` choices (decoupled from beets internals)."
      },
...
      "ImportDecision": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ImportAction"
          },
          "candidate_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Index",
            "description": "Required when `action` is APPLY."
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ImportDecision",
        "description": "The user's answer to a `DecisionRequest` (posted back from the UI)."
      },
      "ImportJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/ImportJobStatus"
          },
          "paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Paths"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "pending_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "decision_submitted": {
            "type": "boolean",
            "title": "Decision Submitted",
            "default": false
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "quiet": {
            "type": "boolean",
            "title": "Quiet",
            "default": false
          },
          "logpath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logpath"
          },
          "group_albums": {
            "type": "boolean",
            "title": "Group Albums",
            "default": false
          },
          "flat": {
            "type": "boolean",
            "title": "Flat",
            "default": false
          },
          "set_fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Set Fields"
          },
          "clean_slate_album_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clean Slate Album Id"
          },
          "clean_slate_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clean Slate Item Id"
          },
          "clean_slate_allow_fewer_files": {
            "type": "boolean",
            "title": "Clean Slate Allow Fewer Files",
            "default": false
          },
          "is_clean_slate": {
            "type": "boolean",
            "title": "Is Clean Slate",
            "description": "Whether this job removes an existing library entry before importing its source afresh.",
            "readOnly": true
          },
          "source_label": {
            "type": "string",
            "title": "Source Label",
            "description": "Display name for the import source: the path basename(s), prefixed for a clean-slate import.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "paths",
          "created_at",
          "is_clean_slate",
          "source_label"
        ],
        "title": "ImportJob",
        "description": "API-facing view of a single import job (the persisted record is `db.models.ImportJobRecord`)."
      },
      "ImportJobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "awaiting_decision",
          "completed",
          "failed",
          "aborted"
        ],
        "title": "ImportJobStatus",
        "description": "Lifecycle states of an import job."
      },
      "ImportSubmitRequest": {
        "properties": {
          "paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Paths",
            "description": "One or more paths (files/dirs) to import."
          },
          "quiet": {
            "type": "boolean",
            "title": "Quiet",
            "description": "Run non-interactively (like `beet import -q`): never prompt. Strong matches are applied automatically; anything else falls back to beets' `quiet_fallback` config (skip by default)."
          },
          "logpath": {
            "anyOf": [
              {
                "type": "string",
                "format": "path"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logpath",
            "description": "Corresponds to the `-l` option when running an import command, or config's `import.log` option. See https://beets.readthedocs.io/en/stable/reference/cli.html#import"
          },
          "group_albums": {
            "type": "boolean",
            "title": "Group Albums",
            "description": "See https://beets.readthedocs.io/en/stable/reference/config.html#group-albums"
          },
          "flat": {
            "type": "boolean",
            "title": "Flat",
            "description": "Corresponds to the `--flat` flag when running an import command. See https://beets.readthedocs.io/en/stable/reference/cli.html#import"
          },
          "set_fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Set Fields",
            "description": "Corresponds to the `--set field=value` option for beets' `import` CLI command. Can set multiple key-value pairs."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paths"
        ],
        "title": "ImportSubmitRequest",
        "description": "Body for starting an import: the filesystem path(s) beets should import. All default values come from the\n`beets.config` values (cached) if left unspecified."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}